(temporarily) updating long commands to show as multiple lines
I need screen shot-able code snippets in order to add them to the Google Forms version of the quiz (which doesn't support code rendering). I'm temporarily chopping up some of the longer commands here so that they are printed to the md as multiple lines. I will change this back momentarily.